% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_minus_one.R
\name{sqrt_minus_one}
\alias{sqrt_minus_one}
\title{square root minus 1}
\usage{
sqrt_minus_one(x)
}
\arguments{
\item{x}{number}
}
\value{
out number
}
\description{
Given a number, this function will take the find the square root minus one
}
\examples{
vals <- rnorm(10)
sqrt_minus_one(vals)
}

