% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_plus_one.R
\name{square_plus_one}
\alias{square_plus_one}
\title{add_one to vector}
\usage{
square_plus_one(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
out a numerical vector
}
\description{
Adds 1 to each element of a vector
}
\examples{
vals <- seq(1, 10, 1)
square_add_one(vals)
}

